Ver3. Fixed services: all processing request move from controllers. added new ViewModel classes. added lazy loading. Fixed IDs - change type to int.

+++ Task9/Service/Implementation/GroupService.cs
@@ -1,52 +1,72 @@
 ﻿using DataAccess;
+using DataAccess.ViewModel;
 using Repository;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Service
 {
     public class GroupService : IGroupService
     {
+        private IRepository<Course> courseRepository;
         private IRepository<Group> groupRepository;
 
-        public GroupService(IRepository<Group> groupRepository)
+        public GroupService(IRepository<Course> courseRepository, IRepository<Group> groupRepository)
         {
+            this.courseRepository = courseRepository;
             this.groupRepository = groupRepository;
         }
 
-        public IEnumerable<Group> GetGroup()
+        public IEnumerable<GroupViewModel> GetGroup()
         {
-            return groupRepository.GetAll();
+            List<GroupViewModel> result = new List<GroupViewModel>();
+            foreach (Group group in groupRepository.GetAll(new Group()))
+            {
+                result.Add(new GroupViewModel(group.Id, group.GroupName, group.Course.Id));
+            }
+            return result;
         }
 
-        public Group GetGroup(long id)
+        public GroupViewModel GetGroup(int id)
         {
-            return groupRepository.Get(id);
+            Group group = groupRepository.Get(id);
+            return new GroupViewModel(group.Id, group.GroupName, group.Course.Id);
         }
 
-        public void InsertGroup(Group group)
+        public void InsertGroup(GroupViewModel groupView)
         {
+            Group group = new Group();
+            group.GroupName = groupView.GroupName;
+            group.Course = courseRepository.Get(groupView.CourseId);
             groupRepository.Insert(group);
         }
 
-        public void UpdateGroup(Group group)
+        public void UpdateGroup(GroupViewModel groupView)
         {
+            Group group = groupRepository.Get(groupView.Id);
+            group.GroupName = groupView.GroupName;
             groupRepository.Update(group);
         }
 
-        public void DeleteGroup(long id)
+        public void DeleteGroup(int id)
         {
-            Group group = GetGroup(id);
+            Group group = groupRepository.Get(id);
+            if (group.Students != null && group.Students.Count() > 0)
+            {
+                throw new ArgumentException("Students list is not empty!");
+            }
             groupRepository.Remove(group);
-            groupRepository.SaveChanges();
         }
 
-        public IEnumerable<Group> GetGroupsByCourse(Course course)
+        public IEnumerable<GroupViewModel> GetGroupsByCourse(int id)
         {
-            return groupRepository.GetAll().Where(item => item.Course == course);
+            List<GroupViewModel> result = new List<GroupViewModel>();
+            foreach (Group group in courseRepository.Get(id).Groups)
+            {
+                result.Add(new GroupViewModel(group.Id, group.GroupName, id));
+            }
+            return result;
         }
     }
 }
